AutoVLA/requirements.txt
torch==2.4.0
torchvision==0.19.0
triton==3.0.0
transformers==4.49.0
transformers-stream-generator==0.0.4
tokenizers==0.21.1
accelerate==1.5.2
peft
qwen-vl-utils==0.0.10
sentencepiece==0.2.1
safetensors==0.5.3
pytorch-lightning==2.2.1
lightning-utilities==0.14.2
torchmetrics==1.7.0
tensorboard
tensorboard-data-server==0.7.2
nuplan-devkit @ git+https://github.com/motional/nuplan-devkit@nuplan-devkit-v1.2
casadi==3.6.7
control==0.9.1
pyquaternion==0.9.9
bokeh==2.4.3
geopandas==1.0.1
shapely==2.0.7
fiona==1.10.1
rasterio==1.3.11
pyogrio==0.10.0
pyproj==3.6.1
rtree==1.4.0
affine==2.4.0
snuggs==1.4.7
opencv-python==4.9.0.80
pillow
timm==1.0.15
av==14.2.0
pandas
scipy
scikit-learn==1.2.2
scikit-image
prettytable
numexpr==2.10.2
datasets==3.6.0
huggingface-hub==0.29.3
pyarrow
lmdb==1.7.5
ray==2.44.1
aioboto3==14.1.0
aiobotocore==2.21.1
boto3==1.37.1
botocore==1.37.1
s3transfer==0.11.3
hydra-core==1.2.0
omegaconf==2.3.0
lm-eval==0.4.9.1
evaluate==0.4.6
nltk==3.9.2
rouge-score==0.1.2
sacrebleu==2.5.1
word2number==1.1
tqdm==4.67.1
tqdm-multiprocess==0.0.11
einops==0.8.1
click==8.1.8
rich==13.9.4
colorama==0.4.6
coloredlogs==15.0.1
humanfriendly==10.0
termcolor==3.1.0
retry==0.9.2
filelock==3.18.0
fsspec==2025.3.0
joblib==1.4.2
jsonlines==4.0.0
portalocker==3.2.0
rootpath==0.1.1
sqlitedict==2.1.0
aiofiles==23.2.1
aiohttp==3.11.14
aioitertools==0.12.0
aiosignal==1.3.2
aiohappyeyeballs==2.6.1
async-timeout==5.0.1
frozenlist==1.5.0
multidict==6.2.0
propcache==0.3.1
yarl==1.18.3
sqlalchemy==1.4.27
msgpack==1.1.0
dill==0.3.8
deepdiff==8.6.1
xxhash==3.6.0
orjson==3.10.16
antlr4-python3-runtime==4.9.3
guppy3==3.1.2
matplotlib
networkx==3.2.1
positional-encodings==6.0.1
protobuf==4.25.3
setuptools
sympy==1.13.3

